+++ internal/presentation/models.go
@@ -11,21 +11,21 @@ type OrderResponse struct {
 	Number     string    `json:"number"`
 	Status     string    `json:"status"`
 	UploadedAt time.Time `json:"uploaded_at"`
-	Accrual    float32   `json:"accrual,omitempty"`
+	Accrual    float64   `json:"accrual,omitempty"`
 }
 
 type BalanceResponse struct {
-	Current  float32 `json:"current"`
-	Withdraw float32 `json:"withdraw"`
+	Current  float64 `json:"current"`
+	Withdraw float64 `json:"withdraw"`
 }
 
 type UploadWithdrawPayload struct {
 	Order string  `json:"order"`
-	Sum   float32 `json:"sum"`
+	Sum   float64 `json:"sum"`
 }
 
 type WithdrawalsResponse struct {
 	Order       string    `json:"order"`
 	ProcessedAt time.Time `json:"processed_at"`
-	Sum         float32   `json:"sum,omitempty"`
+	Sum         float64   `json:"sum,omitempty"`
 }
